hash join + cleanup
