+++ src/daft-local-execution/src/intermediate_ops/intermediate_op.rs
@@ -1,0 +1,10 @@
+use std::sync::Arc;
+
+use common_error::DaftResult;
+use daft_micropartition::MicroPartition;
+
+pub trait IntermediateOperator: dyn_clone::DynClone + Send + Sync {
+    fn execute(&self, input: &[Arc<MicroPartition>]) -> DaftResult<Vec<Arc<MicroPartition>>>;
+}
+
+dyn_clone::clone_trait_object!(IntermediateOperator);
