+++ src/daft-local-execution/src/intermediate_ops/mod.rs
@@ -1,0 +1,5 @@
+pub mod aggregate;
+pub mod fanout;
+pub mod filter;
+pub mod intermediate_op;
+pub mod project;
