+++ src/daft-local-execution/src/sinks/mod.rs
@@ -1,0 +1,5 @@
+pub mod coalesce;
+pub mod hashjoin;
+pub mod limit;
+pub mod reducemerge;
+pub mod sink;
