+++ src/daft-local-execution/src/streaming_pipeline.rs
@@ -1,57 +1,36 @@
-use std::{
-    collections::{BTreeMap, HashMap},
-    sync::Arc,
-};
+use std::{collections::BTreeMap, sync::Arc};
 
 use daft_io::IOStatsContext;
 use daft_micropartition::MicroPartition;
-use daft_plan::{
-    physical_ops::{
-        Aggregate, Coalesce, Filter, HashJoin, InMemoryScan, Limit, Project, TabularScan,
-    },
-    PhysicalPlan,
-};
-use daft_scan::ScanTask;
 use tokio::sync::{mpsc, Mutex};
 
 use crate::{
-    intermediate_op::IntermediateOperatorType,
-    sink::{CoalesceSink, HashJoinSink, LimitSink, Sink, SinkResultType},
+    intermediate_ops::intermediate_op::IntermediateOperator,
+    sinks::sink::{Sink, SinkResultType},
+    source::{Morsel, Source},
 };
-
-enum Source {
-    Data(Vec<Arc<MicroPartition>>),
-    ScanTask(Vec<Arc<ScanTask>>),
-    Receiver(mpsc::Receiver<Arc<MicroPartition>>),
-}
-
-enum Morsel {
-    Data(Arc<MicroPartition>),
-    ScanTask(Arc<ScanTask>),
-}
 
 struct Producer {
     id: usize,
     input: Source,
-    final_tx: mpsc::Sender<(usize, usize, Arc<MicroPartition>)>,
-    intermediate_operators: Vec<IntermediateOperatorType>,
+    final_tx: mpsc::Sender<(usize, usize, Vec<Arc<MicroPartition>>)>,
+    intermediate_operators: Vec<Box<dyn IntermediateOperator>>,
 }
 
 impl Producer {
     async fn run(self) {
         let mut handles = vec![];
-        // TODO: Factor out all the dupes
         match self.input {
             Source::Data(data) => {
                 for value in data {
                     let final_tx = self.final_tx.clone();
                     let part_idx = handles.len();
 
-                    // Spawn a consumer for each piece of data
+                    // Spawn a consumer for each morsel
                     let handle = tokio::spawn(
                         Consumer::new(
                             self.id,
-                            Morsel::Data(value),
+                            Morsel::Data(vec![value]),
                             final_tx,
                             part_idx,
                             self.intermediate_operators.clone(),
@@ -66,7 +45,7 @@ impl Producer {
                     let final_tx = self.final_tx.clone();
                     let id = handles.len();
 
-                    // Spawn a consumer for each piece of data
+                    // Spawn a consumer for each morsel
                     let handle = tokio::spawn(
                         Consumer::new(
                             self.id,
@@ -85,7 +64,7 @@ impl Producer {
                     let final_tx = self.final_tx.clone();
                     let part_idx = handles.len();
 
-                    // Spawn a consumer for each piece of data
+                    // Spawn a consumer for each morsel
                     let handle = tokio::spawn(
                         Consumer::new(
                             self.id,
@@ -111,18 +90,18 @@ impl Producer {
 struct Consumer {
     id: usize,
     input: Morsel,
-    tx: mpsc::Sender<(usize, usize, Arc<MicroPartition>)>,
+    tx: mpsc::Sender<(usize, usize, Vec<Arc<MicroPartition>>)>,
     part_idx: usize,
-    intermediate_operators: Vec<IntermediateOperatorType>,
+    intermediate_operators: Vec<Box<dyn IntermediateOperator>>,
 }
 
 impl Consumer {
     fn new(
         id: usize,
         input: Morsel,
-        tx: mpsc::Sender<(usize, usize, Arc<MicroPartition>)>,
+        tx: mpsc::Sender<(usize, usize, Vec<Arc<MicroPartition>>)>,
         part_idx: usize,
-        intermediate_operators: Vec<IntermediateOperatorType>,
+        intermediate_operators: Vec<Box<dyn IntermediateOperator>>,
     ) -> Self {
         Self {
             id,
@@ -135,16 +114,20 @@ impl Consumer {
 
     async fn run(self) {
         match self.input {
-            Morsel::Data(mut value) => {
+            Morsel::Data(value) => {
                 // Apply intermediate operators
+                let mut results = value;
                 for intermediate_operator in &self.intermediate_operators {
-                    value = intermediate_operator.execute(&value).unwrap();
+                    results = intermediate_operator.execute(&results).unwrap();
                 }
                 // Send the value to the final consumer
-                let _ = self.tx.send((self.id, self.part_idx, value)).await;
+                let _ = self.tx.send((self.id, self.part_idx, results)).await;
             }
             Morsel::ScanTask(scan_task) => {
                 let (send, recv) = tokio::sync::oneshot::channel();
+                // TODO: need to put the scan task in a separate thread because it tries to create another tokio runtime,
+                // which is not allowed because it is already in a tokio runtime
+                // rayon is just a patch, need to think about a better solution
                 rayon::spawn(move || {
                     let io_stats = IOStatsContext::new(format!(
                         "MicroPartition::from_scan_task for {:?}",
@@ -154,28 +137,33 @@ impl Consumer {
                         Arc::new(MicroPartition::from_scan_task(scan_task, io_stats).unwrap());
                     let _ = send.send(part);
                 });
-                let mut value = recv.await.unwrap();
+                let value = recv.await.unwrap();
+
                 // Apply intermediate operators
+                let mut results = vec![value];
                 for intermediate_operator in &self.intermediate_operators {
-                    value = intermediate_operator.execute(&value).unwrap();
+                    results = intermediate_operator.execute(&results).unwrap();
                 }
                 // Send the value to the final consumer
-                self.tx.send((self.id, self.part_idx, value)).await.unwrap();
+                self.tx
+                    .send((self.id, self.part_idx, results))
+                    .await
+                    .unwrap();
             }
         }
     }
 }
 
 struct FinalConsumer {
-    rx: mpsc::Receiver<(usize, usize, Arc<MicroPartition>)>,
-    next_pipeline_tx: Option<mpsc::Sender<Arc<MicroPartition>>>,
+    rx: mpsc::Receiver<(usize, usize, Vec<Arc<MicroPartition>>)>,
+    next_pipeline_tx: Option<mpsc::Sender<Vec<Arc<MicroPartition>>>>,
     sink: Option<Arc<Mutex<dyn Sink>>>,
 }
 
 impl FinalConsumer {
     fn new(
-        rx: mpsc::Receiver<(usize, usize, Arc<MicroPartition>)>,
-        next_pipeline_tx: Option<mpsc::Sender<Arc<MicroPartition>>>,
+        rx: mpsc::Receiver<(usize, usize, Vec<Arc<MicroPartition>>)>,
+        next_pipeline_tx: Option<mpsc::Sender<Vec<Arc<MicroPartition>>>>,
         sink: Option<Arc<Mutex<dyn Sink>>>,
     ) -> Self {
         Self {
@@ -191,6 +179,8 @@ impl FinalConsumer {
             Some(ref sink) => sink.lock().await.in_order(),
             None => false,
         };
+        let mut done = false;
+
         if in_order {
             // Buffer values until they can be processed in order
             let mut buffer = BTreeMap::new();
@@ -200,23 +190,25 @@ impl FinalConsumer {
                 buffer.insert(part_idx, (id, value));
 
                 // Process all consecutive values starting from `next_index`
-                while let Some((id, val)) = buffer.get(&next_index).as_ref() {
+                while let Some((id, val)) = buffer.remove(&next_index) {
+                    next_index += 1;
                     if let Some(sink) = self.sink.as_mut() {
                         // Sink the value
-                        let sink_result = sink.lock().await.sink(val, *id);
+                        let sink_result = sink.lock().await.sink(&val, id);
                         match sink_result {
                             SinkResultType::Finished(part) => {
                                 if let Some(part) = part {
                                     if let Some(ref tx) = self.next_pipeline_tx {
-                                        tx.send(part).await.unwrap();
+                                        tx.send(vec![part]).await.unwrap();
                                     }
                                 }
+                                done = true;
                                 break;
                             }
                             SinkResultType::NeedMoreInput(part) => {
                                 if let Some(part) = part {
                                     if let Some(ref tx) = self.next_pipeline_tx {
-                                        tx.send(part).await.unwrap();
+                                        tx.send(vec![part]).await.unwrap();
                                     }
                                 }
                             }
@@ -227,39 +219,41 @@ impl FinalConsumer {
                             tx.send(val.clone()).await.unwrap();
                         }
                     }
-                    buffer.remove(&next_index);
-                    next_index += 1;
                 }
             }
 
             // Process remaining values
-            while let Some((id, val)) = buffer.remove(&next_index) {
-                if let Some(sink) = self.sink.as_mut() {
-                    // Sink the value
-                    let sink_result = sink.lock().await.sink(&val, id);
-                    match sink_result {
-                        SinkResultType::Finished(part) => {
-                            if let Some(part) = part {
-                                if let Some(ref tx) = self.next_pipeline_tx {
-                                    tx.send(part).await.unwrap();
+            if !done {
+                while let Some((id, val)) = buffer.remove(&next_index) {
+                    if let Some(sink) = self.sink.as_mut() {
+                        // Sink the value
+                        let sink_result = sink.lock().await.sink(&val, id);
+                        match sink_result {
+                            SinkResultType::Finished(part) => {
+                                if let Some(part) = part {
+                                    if let Some(ref tx) = self.next_pipeline_tx {
+                                        tx.send(vec![part]).await.unwrap();
+                                    }
+                                }
+                                done = true;
+                                break;
+                            }
+                            SinkResultType::NeedMoreInput(part) => {
+                                if let Some(part) = part {
+                                    if let Some(ref tx) = self.next_pipeline_tx {
+                                        tx.send(vec![part]).await.unwrap();
+                                    }
                                 }
                             }
                         }
-                        SinkResultType::NeedMoreInput(part) => {
-                            if let Some(part) = part {
-                                if let Some(ref tx) = self.next_pipeline_tx {
-                                    tx.send(part).await.unwrap();
-                                }
-                            }
+                    } else {
+                        // If there is no sink, just forward the value
+                        if let Some(ref tx) = self.next_pipeline_tx {
+                            tx.send(val.clone()).await.unwrap();
                         }
                     }
-                } else {
-                    // If there is no sink, just forward the value
-                    if let Some(ref tx) = self.next_pipeline_tx {
-                        tx.send(val.clone()).await.unwrap();
-                    }
+                    next_index += 1;
                 }
-                next_index += 1;
             }
         } else {
             while let Some((id, _, val)) = self.rx.recv().await {
@@ -270,14 +264,16 @@ impl FinalConsumer {
                         SinkResultType::Finished(part) => {
                             if let Some(part) = part {
                                 if let Some(ref tx) = self.next_pipeline_tx {
-                                    tx.send(part).await.unwrap();
+                                    tx.send(vec![part]).await.unwrap();
                                 }
                             }
+                            done = true;
+                            break;
                         }
                         SinkResultType::NeedMoreInput(part) => {
                             if let Some(part) = part {
                                 if let Some(ref tx) = self.next_pipeline_tx {
-                                    tx.send(part).await.unwrap();
+                                    tx.send(vec![part]).await.unwrap();
                                 }
                             }
                         }
@@ -291,12 +287,13 @@ impl FinalConsumer {
             }
         }
 
-        // TODO: Probs need to flush the sink here
-        if let Some(sink) = self.sink {
-            if let Some(leftover) = sink.lock().await.finalize() {
-                for part in leftover {
-                    if let Some(ref tx) = self.next_pipeline_tx {
-                        tx.send(part).await.unwrap();
+        if !done {
+            if let Some(sink) = self.sink {
+                if let Some(leftover) = sink.lock().await.finalize() {
+                    for part in leftover {
+                        if let Some(ref tx) = self.next_pipeline_tx {
+                            tx.send(vec![part]).await.unwrap();
+                        }
                     }
                 }
             }
@@ -307,16 +304,16 @@ impl FinalConsumer {
 pub struct StreamingPipeline {
     id: usize,
     input: Source,
-    next_pipeline_tx: Option<mpsc::Sender<Arc<MicroPartition>>>,
-    intermediate_operators: Vec<IntermediateOperatorType>,
+    next_pipeline_tx: Option<mpsc::Sender<Vec<Arc<MicroPartition>>>>,
+    intermediate_operators: Vec<Box<dyn IntermediateOperator>>,
     sink: Option<Arc<Mutex<dyn Sink>>>,
 }
 
 impl StreamingPipeline {
-    fn new(
+    pub fn new(
         id: usize,
         input: Source,
-        next_pipeline_tx: Option<mpsc::Sender<Arc<MicroPartition>>>,
+        next_pipeline_tx: Option<mpsc::Sender<Vec<Arc<MicroPartition>>>>,
     ) -> Self {
         Self {
             id,
@@ -327,15 +324,15 @@ impl StreamingPipeline {
         }
     }
 
-    fn add_operator(&mut self, operator: IntermediateOperatorType) {
+    pub fn add_operator(&mut self, operator: Box<dyn IntermediateOperator>) {
         self.intermediate_operators.push(operator);
     }
 
-    fn set_sink(&mut self, sink: Arc<Mutex<dyn Sink>>) {
+    pub fn set_sink(&mut self, sink: Arc<Mutex<dyn Sink>>) {
         self.sink = Some(sink);
     }
 
-    pub fn set_next_pipeline_tx(&mut self, tx: mpsc::Sender<Arc<MicroPartition>>) {
+    pub fn set_next_pipeline_tx(&mut self, tx: mpsc::Sender<Vec<Arc<MicroPartition>>>) {
         self.next_pipeline_tx = Some(tx);
     }
 
@@ -363,147 +360,4 @@ impl StreamingPipeline {
         // Await the final consumer
         final_handle.await.unwrap();
     }
-}
-
-pub fn physical_plan_to_streaming_pipeline(
-    physical_plan: &Arc<PhysicalPlan>,
-    psets: &HashMap<String, Vec<Arc<MicroPartition>>>,
-) -> Vec<StreamingPipeline> {
-    let mut pipelines = vec![];
-    fn recursively_create_pipelines<'a>(
-        physical_plan: &Arc<PhysicalPlan>,
-        psets: &HashMap<String, Vec<Arc<MicroPartition>>>,
-        pipelines: &'a mut Vec<StreamingPipeline>,
-    ) -> &'a mut StreamingPipeline {
-        match physical_plan.as_ref() {
-            PhysicalPlan::InMemoryScan(InMemoryScan { in_memory_info, .. }) => {
-                let partitions = psets.get(&in_memory_info.cache_key).unwrap();
-                let new_pipeline =
-                    StreamingPipeline::new(pipelines.len(), Source::Data(partitions.clone()), None);
-                pipelines.push(new_pipeline);
-                pipelines.last_mut().unwrap()
-            }
-            PhysicalPlan::TabularScan(TabularScan { scan_tasks, .. }) => {
-                let new_pipeline = StreamingPipeline::new(
-                    pipelines.len(),
-                    Source::ScanTask(scan_tasks.clone()),
-                    None,
-                );
-                pipelines.push(new_pipeline);
-                pipelines.last_mut().unwrap()
-            }
-            PhysicalPlan::Aggregate(Aggregate {
-                input,
-                aggregations,
-                groupby,
-            }) => {
-                let current_pipeline = recursively_create_pipelines(input, psets, pipelines);
-                let agg_op = IntermediateOperatorType::Aggregate {
-                    aggregations: aggregations.clone(),
-                    groupby: groupby.clone(),
-                };
-                current_pipeline.add_operator(agg_op);
-                current_pipeline
-            }
-            PhysicalPlan::Project(Project {
-                input, projection, ..
-            }) => {
-                let current_pipeline = recursively_create_pipelines(input, psets, pipelines);
-                let proj_op = IntermediateOperatorType::Project {
-                    projection: projection.clone(),
-                };
-                current_pipeline.add_operator(proj_op);
-                current_pipeline
-            }
-            PhysicalPlan::Filter(Filter { input, predicate }) => {
-                let current_pipeline = recursively_create_pipelines(input, psets, pipelines);
-                let filter_op = IntermediateOperatorType::Filter {
-                    predicate: predicate.clone(),
-                };
-                current_pipeline.add_operator(filter_op);
-                current_pipeline
-            }
-            PhysicalPlan::Limit(Limit { limit, input, .. }) => {
-                let current_pipeline = recursively_create_pipelines(input, psets, pipelines);
-                let sink = LimitSink::new(*limit as usize);
-                current_pipeline.set_sink(Arc::new(Mutex::new(sink)));
-
-                let (tx, rx) = tokio::sync::mpsc::channel::<Arc<MicroPartition>>(32);
-                current_pipeline.set_next_pipeline_tx(tx);
-
-                let new_pipeline =
-                    StreamingPipeline::new(pipelines.len(), Source::Receiver(rx), None);
-                pipelines.push(new_pipeline);
-                pipelines.last_mut().unwrap()
-            }
-            PhysicalPlan::Coalesce(Coalesce {
-                input,
-                num_from,
-                num_to,
-            }) => {
-                let current_pipeline = recursively_create_pipelines(input, psets, pipelines);
-
-                let q = num_from / num_to;
-                let r = num_from % num_to;
-
-                let mut distribution = vec![q; *num_to];
-                for bucket in distribution.iter_mut().take(r) {
-                    *bucket += 1;
-                }
-                let sink = CoalesceSink::new(distribution);
-
-                current_pipeline.set_sink(Arc::new(Mutex::new(sink)));
-
-                let (tx, rx) = tokio::sync::mpsc::channel::<Arc<MicroPartition>>(32);
-                current_pipeline.set_next_pipeline_tx(tx);
-
-                let new_pipeline =
-                    StreamingPipeline::new(pipelines.len(), Source::Receiver(rx), None);
-                pipelines.push(new_pipeline);
-                pipelines.last_mut().unwrap()
-            }
-            PhysicalPlan::HashJoin(HashJoin {
-                left,
-                right,
-                left_on,
-                right_on,
-                join_type,
-            }) => {
-                {
-                    recursively_create_pipelines(left, psets, pipelines);
-                };
-                let left_idx = pipelines.len() - 1;
-                {
-                    recursively_create_pipelines(right, psets, pipelines);
-                };
-                let right_idx = pipelines.len() - 1;
-
-                let hash_join_sink = HashJoinSink::new(
-                    left_on.clone(),
-                    right_on.clone(),
-                    *join_type,
-                    left_idx,
-                    right_idx,
-                );
-                let ptr = Arc::new(Mutex::new(hash_join_sink));
-                let (tx, rx) = tokio::sync::mpsc::channel::<Arc<MicroPartition>>(32);
-
-                pipelines[left_idx].set_sink(ptr.clone());
-                pipelines[right_idx].set_sink(ptr);
-                pipelines[left_idx].set_next_pipeline_tx(tx.clone());
-                pipelines[right_idx].set_next_pipeline_tx(tx);
-
-                let new_pipeline =
-                    StreamingPipeline::new(pipelines.len(), Source::Receiver(rx), None);
-                pipelines.push(new_pipeline);
-                pipelines.last_mut().unwrap()
-            }
-            _ => {
-                todo!()
-            }
-        }
-    }
-
-    recursively_create_pipelines(physical_plan, psets, &mut pipelines);
-    pipelines
 }
